Separate different types of tests into separate files
